pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-flow"
version = "0.1.0"
description = "CLI tool to make project folders AI-ready for VS Code, Cursor, Kilo, and more."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "Jinja2>=3.1",
    "python-frontmatter>=1.1",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
project-flow = "project_flow.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
